Aritmetical ioerations

+++ Exam Preparation/2 - Arithmetical operations.cpp
@@ -7,11 +7,10 @@ using namespace std;
 
 int main ()
 {
-    // Be careful when both operands are integers!
+    // Be careful when both operands are integers! (// in Python)
     int x = 10;
     int y = 3;
-
-    cout << "X / Y = " << x/y << endl;
+    cout << "X / Y = " << (double) x/y << endl;
 
     return 0;
 }
